Refactor: DTO 클래스 이름 수정 및 이동
